+++ src/OData/OperationCenter.cs
@@ -1,75 +1,17 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Globalization;
-using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using SenseNet.ApplicationModel;
 using SenseNet.ContentRepository;
-using SenseNet.ContentRepository.Fields;
-using SenseNet.ContentRepository.Schema;
-using SenseNet.ContentRepository.Storage;
 using SenseNet.ContentRepository.Storage.Security;
 
 namespace SenseNet.OData
 {
-    //UNDONE: one class per file.
-    internal class OperationMethodStorage : IOperationMethodStorage
-    {
-        public IEnumerable<ActionBase> GetActions(IEnumerable<ActionBase> storedActions, Content content, string scenario)
-        {
-            // Local cache
-            string[] actualRoles = null;
-            // Gets role names of the current user and uses the local cache
-            string[] GetRoles()
-            {
-                if (actualRoles == null)
-                    actualRoles = NodeHead.Get(SecurityHandler.GetGroups()).Select(y => y.Name).ToArray();
-                return actualRoles;
-            }
-
-            var stored = storedActions.ToArray();
-            var operationMethods = OperationCenter.Operations
-                .SelectMany(x => x.Value)
-                .Where(x => FilterByApplications(x.Method.Name, stored))
-                .Where(x => FilterByContentTypes(content.ContentType, x.ContentTypes))
-                .Where(x => x.Roles.Length == 0 || GetRoles().Intersect(x.Roles).Any())
-                .Select(x => new ODataOperationMethodAction(x));
-
-            return stored.Union(operationMethods).ToArray();
-        }
-
-        private bool FilterByApplications(string operationName, ActionBase[] stored)
-        {
-            for (int i = 0; i < stored.Length; i++)
-                if (stored[i].Name == operationName)
-                    return false;
-            return true;
-        }
-        private bool FilterByContentTypes(ContentType contentType, string[] allowedContentTypeNames)
-        {
-            if (allowedContentTypeNames.Length == 0)
-                return true;
-
-            for (int i = 0; i < allowedContentTypeNames.Length; i++)
-            {
-                var existingContentType = ContentType.GetByName(allowedContentTypeNames[i]);
-                if (existingContentType == null)
-                    continue;
-                if (contentType.IsInstaceOfOrDerivedFrom(existingContentType.Name))
-                    return true;
-            }
-
-            return false;
-        }
-
-    }
-
     public class OperationCenter
     {
         private static readonly OperationInfo[] EmptyMethods = new OperationInfo[0];
@@ -200,7 +142,7 @@ namespace SenseNet.OData
                 candidates = candidates.Where(x =>
                         AllRequiredParametersExist(x, requestParameterNames) &&
                         FilterByContentTypes(x.ContentTypes, content.ContentType.Name) &&
-                        FilterByRolesAndPermissions(x.Roles, x.RequiredPermissions, content, User.Current))
+                        FilterByRolesAndPermissions(x.Roles, x.Permissions, content, User.Current))
                     .ToArray();
             }
